+++ src/cli/repo/prompt.rs
@@ -50,7 +50,7 @@ pub async fn run(config: &Config, args: PromptArgs) -> i32 {
     };
 
     let mut prompt = Prompt::new(&repository);
-    let ret = repository.get_vcs_repo().prompt(config, &mut prompt).await;
+    let ret = repository.get_vcs_repo().prompt(config, &mut prompt);
     if ret == 0 {
         println!("{}", prompt.display(config));
     }